Car profile fix

+++ Frontend/src/Components/Profile/AddPhotoToGallery/AddPhotoToProfileGallery.tsx
@@ -2,17 +2,19 @@ import { useState, useEffect } from 'react'
 import "./AddPhotoToProfileGallery.css"
 import { storage } from "../../../base";
 import axios, { AxiosError } from 'axios';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import { getUser } from '../../../Store/authSlice';
-import { Car } from '../../../Store/carSlice';
+import { Car, loadCarsData } from '../../../Store/carSlice';
+import { AppDispatch } from '../../../Store/store';
 
 const AddPhotoToProfileGallery = () => {
     const [titleValue, setTitleValue] = useState('')
     const [currentFile, setCurrentFile] = useState<Blob | null>(null)
+    const dispatch = useDispatch<AppDispatch>()
     const user = useSelector(getUser)
     const car = useSelector(Car)
     const [carId, setCarId] = useState(car?._id || "")
-    const [filePath, setFilePath] = useState<string>(user?.uid + "/" + carId + "/")
+    const [filePath, setFilePath] = useState<string>(user?.uid + "/" + car?._id || "" + "/")
     const [imageSource, setImageSource] = useState<string>(car?.profileUrl || "")
 
     const handleProfileImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -23,6 +25,12 @@ const AddPhotoToProfileGallery = () => {
             setTitleValue(file.name)
         }
     };
+
+    useEffect(() => {
+      setCarId(car?._id || "")
+      setFilePath(user?.uid + "/" + car?._id || "" + "/")
+      setImageSource(car?.profileUrl || "")
+    }, [car])
 
     useEffect(() => {
         if(currentFile) {
@@ -37,7 +45,6 @@ const AddPhotoToProfileGallery = () => {
         let fullUrl = ""
         const storageRef = storage.ref();
         let name = filePath + titleValue;
-        console.log(name)
         //let name = user?.uid + "/gallery/" + titleValue;//////////////////////
         const fileRef = storageRef.child(name);
         try {
@@ -51,8 +58,11 @@ const AddPhotoToProfileGallery = () => {
         const postData = {
             name: name,
             title: titleValue,
-            fullUrl: fullUrl
+            fullUrl: fullUrl,
+            profile: true,
+            carId: car?._id
           };
+          console.log(postData)
           await axios.post("http://127.0.0.1:3000/post_photo_to_gallery", postData, {
             headers: {
               "Content-Type": "application/json",
@@ -62,15 +72,15 @@ const AddPhotoToProfileGallery = () => {
             console.error("Wystąpił błąd:", error);
           });
         let profilemedia = car?.media.find(car => car.profile === true)
-        await axios.delete("http://127.0.0.1:3000/delete_photo", { data: { _id: profilemedia?._id }})
+        if(profilemedia)
+          await axios.delete("http://127.0.0.1:3000/delete_photo", { data: { _id: profilemedia?._id }})
         
         setTitleValue("")
-        updateCarImageProfile(name)
         setCurrentFile(null);
+        await dispatch(loadCarsData())
     }
 
-    const updateCarImageProfile = (name: string) => {
-        console.log(name)
+    /*const updateCarImageProfile = (name: string) => {
         axios.put("http://127.0.0.1:3000/updateCarMedia", { image: name, carId: carId, profile: true }, {
             headers: {
               "Content-Type": "application/json",
@@ -79,7 +89,7 @@ const AddPhotoToProfileGallery = () => {
           .catch((error: AxiosError) => {
             console.error("Wystąpił błąd:", error);
           })
-    }
+    }*/
 
     const isImage = () => { return imageSource != '' && imageSource != "domyślny_url_obrazka" }
